requirements.txt
absl-py==0.11.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
dill==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
idna==2.10
importlib-metadata==3.1.0
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.2
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.1.2.30
pandas==1.1.4
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.50.2
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0
